testsapp/requirements.txt
pytest
pytest-django
pytest-cov
boto
south